+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 🛑 Breaking changes 🛑
+
+- Remove `reconnection_delay` from OpenCensus exporter #1516.
+
 ## v0.8.0 Beta
 
 ## 🚀 New components 🚀